[CHANGE] Renamed template tag library
