+++ aasrp/templates/aasrp/ajax_render/srp_request_additional_information.html
@@ -1,4 +1,4 @@
-{% load aasrp_user %}
+{% load aasrp %}
 {% load humanize %}
 {% load i18n %}
