+++ aasrp/templates/aasrp/bundles/aa-srp-dashboard-js.html
@@ -1,3 +1,3 @@
-{% load aasrp_versioned_static %}
+{% load aasrp %}
 
 <script src="{% aasrp_static 'aasrp/javascript/aa-srp-dashboard.min.js' %}"></script>
