+++ aasrp/templates/aasrp/bundles/aa-srp-form-css.html
@@ -1,3 +1,3 @@
-{% load aasrp_versioned_static %}
+{% load aasrp %}
 
 <link rel="stylesheet" href="{% aasrp_static 'aasrp/css/aa-srp-form.min.css' %}">
