+++ aasrp/templates/aasrp/bundles/aa-srp-view-requests-js.html
@@ -1,3 +1,3 @@
-{% load aasrp_versioned_static %}
+{% load aasrp %}
 
 <script src="{% aasrp_static 'aasrp/javascript/aa-srp-view-requests.min.js' %}"></script>
